moved utilities to separate directory
